assignment/websrv/webpages/csDashboard/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
click==8.0.4
command-not-found==0.3
distlib==0.3.4
Django==4.0.2
filelock==3.6.0
Flask==2.0.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
python-apt==2.0.0+ubuntu0.20.4.7
six==1.16.0
sqlparse==0.4.2
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
virtualenv==20.13.2
Werkzeug==2.0.3
